feat: implement create review feature

+++ app/src/main/java/com/example/toyproject5/ui/screens/ParticipantsScreen.kt
@@ -102,7 +102,7 @@ fun ParticipantsScreen(groupId: Int, groupName: String, onBack: () -> Unit, navC
                             email = participant.username,
                             onItemClick = {
                                 // 참여자의 고유 ID를 가지고 프로필 화면으로 이동!
-                                navController.navigate(NavRoute.UserProfile.createRoute(participant.userId.toInt()))
+                                navController.navigate(NavRoute.UserProfile.createRoute(participant.userId))
                             }
                         )
                     }
